backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
anthropic>=0.36.0
python-frontmatter>=1.1.0
python-dotenv>=1.0.0
